refactor: change calling function toggle classes
