+++ src/js/nav.js
@@ -1,3 +1,9 @@
-export default navigation = () => {
-    document.querySelector('.navigation_button').toggle('change')
+export const toggleClassesNavigation = () => {
+    const navList = document.querySelector('.navigation_list');
+    const navButton = document.querySelector('.navigation_button');
+    
+    navButton.addEventListener('click', () => {
+        navButton.classList.toggle('change');
+        navList.classList.toggle('active');
+    })
 };
